refactor(account): 重构账户相关代码
- 修改 Account 实体类，将 name 字段改为 memberName
- 更新 AccountMapper 接口和 XML 文件中的 SQL 语句
- 调整 AccountServiceImpl 中的分页查询逻辑
- 重构 AccountVO 类，调整字段名称和结构
- 新增 MemberException 异常类
- 修改 MemberServiceImpl 中的删除和更新逻辑
- 在 MessageConstant 中添加新的错误提示信息

+++ budget-server/src/main/java/com/budget/service/impl/MemberServiceImpl.java
@@ -82,8 +82,13 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public void deleteMemberById(Long id) {
+        //不能删除管理员
         if(id.equals(AdminConstant.ADMIN_ID)){
             throw  new AdminException(MessageConstant.ADMIN_DELETE_NOT_ALLOWED);
+        }
+        //用户不能删除自己
+        if(id.equals(BaseContext.getCurrentId())){
+            throw  new MemberException(MessageConstant.MEMBER_DELETE_NOT_ALLOWED);
         }
         memberMapper.deleteById(id);
     }
@@ -95,7 +100,11 @@ public class MemberServiceImpl implements MemberService {
         }
         Member member = new Member();
         BeanUtils.copyProperties(memberDTO, member);
-
+        String password = member.getPassword();
+        if(password !=null){
+            String md5Password=DigestUtils.md5DigestAsHex(password.getBytes());
+            member.setPassword(md5Password);
+        }
         memberMapper.update(member);
     }
